rename hook to useSupportedRoutes
